requirements.txt
numpy==1.19.2
streamlit==0.71.0
opencv-python==4.4.0.46
Pillow==2.2.2